+++ message.go
@@ -9,10 +9,12 @@ import (
 // to the RespondFunc that was registered via Bot.Respond(…) or Bot.RespondRegex(…)
 // when the message matches the regular expression of the handler.
 type Message struct {
-	Context context.Context
-	Text    string
-	Channel string
-	Matches []string // contains all sub matches of the regular expression that matched the Text
+	Context  context.Context
+	Text     string
+	AuthorID string
+	Channel  string
+	Matches  []string    // contains all sub matches of the regular expression that matched the Text
+	Data     interface{} // corresponds to the ReceiveMessageEvent.Data field
 
 	adapter Adapter
 }